BUG fix on SQL

+++ dbmodel/ws/ws_53b_rtc_views.sql
@@ -42,7 +42,7 @@ connec.code as urban_propierties_code,
 
 FROM rtc_hydrometer
 LEFT JOIN ext_rtc_hydrometer ON ext_rtc_hydrometer.hydrometer_id::integer = rtc_hydrometer.hydrometer_id::integer
-LEFT JOIN ext_cat_hydrometer ON ext_cat_hydrometer.id::integer = ext_rtc_hydrometer.cat_hydrometer_id::integer
+LEFT JOIN ext_cat_hydrometer ON ext_cat_hydrometer.id = ext_rtc_hydrometer.cat_hydrometer_id
 JOIN rtc_hydrometer_x_connec ON rtc_hydrometer_x_connec.hydrometer_id::integer = rtc_hydrometer.hydrometer_id::integer
 JOIN  connec ON rtc_hydrometer_x_connec.connec_id=connec.connec_id;
